Move components and views into their own dirs
