+++ beta/components/albums/index.js
@@ -5,7 +5,7 @@ const Playlist = require('@resonate/playlist-component')
 const adapter = require('@resonate/schemas/adapters/v1/track')
 const nanostate = require('nanostate')
 const nanologger = require('nanologger')
-const Loader = require('./play-count')
+const Loader = require('../play-count')
 const clock = require('mm-ss')
 const compare = require('nanocomponent/compare')
 const Pagination = require('@resonate/pagination')
