+++ beta/components/artists/index.js
@@ -1,51 +1,18 @@
-const Component = require('choo/component')
+const Nanocomponent = require('choo/component')
 const compare = require('nanocomponent/compare')
 const html = require('choo/html')
 const clone = require('shallow-clone')
+const Artist = require('./single')
 const nanostate = require('nanostate')
 const nanologger = require('nanologger')
-const Loader = require('./play-count')
+const Loader = require('../play-count')
 const Pagination = require('@resonate/pagination')
-const css = require('sheetify')
-const prefix = css`
-  @custom-media --breakpoint-not-small screen and (min-width: 30em);
-  @custom-media --breakpoint-medium screen and (min-width: 30em) and (max-width: 60em);
-  @custom-media --breakpoint-large screen and (min-width: 60em);
 
-  @media(--breakpoint-large) {
-    :host:first-child {
-      width: 40%;
-      margin-bottom: -1px;
-    }
-  }
-`
-
-class LabelItem extends Component {
-  createElement (props) {
-    const { avatar: image = {}, id, name } = props
-    const fallback = image.original || '/assets/default.png'
-    const { large: imageUrl = fallback } = image
-
-    return html`
-      <li class="${prefix} fl w-50 w-third-m w-20-l pa3 grow">
-        <a class="db aspect-ratio aspect-ratio--1x1 bg-dark-gray" href="/labels/${id}">
-          <img aria-label=${name} src=${imageUrl} decoding="auto" class="aspect-ratio--object">
-          <span class="absolute bottom-0 truncate w-100 h2" style="top:100%;">
-            ${name}
-          </span>
-        </a>
-      </li>
-    `
-  }
-
-  update () {
-    return false
-  }
-}
-
-class Labels extends Component {
+class Artists extends Nanocomponent {
   constructor (id, state, emit) {
     super(id)
+
+    this.items = []
 
     this.state = state
     this.emit = emit
@@ -53,9 +20,7 @@ class Labels extends Component {
 
     this.log = nanologger(id)
 
-    this.items = []
-
-    this.renderLabels = this.renderLabels.bind(this)
+    this.renderArtists = this.renderArtists.bind(this)
     this.renderError = this.renderError.bind(this)
     this.renderPlaceholder = this.renderPlaceholder.bind(this)
 
@@ -77,79 +42,84 @@ class Labels extends Component {
     })
 
     this.local.loader.on('loader:toggle', () => {
-      this.log.info('loader:toggle', this.local.loader.state.loader)
       if (this.element) this.rerender()
     })
 
     this.local.machine.on('notFound', () => {
-      this.log.info('notFound')
       if (this.element) this.rerender()
-    })
-
-    this.local.machine.on('loading', () => {
-      this.log.info('loading')
     })
 
     this.local.machine.on('error', () => {
-      this.log.error('error')
-      if (this.element) this.rerender()
-    })
-
-    this.local.machine.on('data', () => {
-      this.log.info('data')
       if (this.element) this.rerender()
     })
   }
 
   createElement (props = {}) {
     const self = this
-    const { items = [], pagination: paginationEnabled = true, numberOfPages = 1 } = props
+    const { numberOfPages = 1, pagination: paginationEnabled = true } = props
 
-    this.items = clone(items)
+    this.local.shuffle = props.shuffle
+    this.local.items = props.items || []
 
-    const labels = {
+    const artists = {
       loading: {
         on: this.renderLoader,
         off: () => {}
       }[this.local.loader.state.loader](),
       notFound: this.renderPlaceholder(),
       error: this.renderError()
-    }[this.local.machine.state] || this.renderLabels()
+    }[this.local.machine.state] || this.renderArtists()
 
     let paginationEl
 
-    if (paginationEnabled && numberOfPages > 1) {
-      paginationEl = new Pagination('labels-pagination', this.state, this.emit).render({
+    if (paginationEnabled) {
+      paginationEl = new Pagination('artists-pagination', this.state, this.emit).render({
         navigate: function (pageNumber) {
-          self.emit(self.state.events.PUSHSTATE, self.state.href + `?page=${pageNumber}`)
+          const path = !/artists/.test(this.state.href) ? '/artists' : ''
+          self.emit(self.state.events.PUSHSTATE, self.state.href + `${path}?page=${pageNumber}`)
         },
+        path: !/artists/.test(this.state.href) ? '/artists' : '',
         numberOfPages
       })
     }
 
     return html`
       <div class="flex flex-column flex-auto w-100">
-        ${labels}
+        ${artists}
         ${paginationEl}
       </div>
     `
   }
 
-  renderLabels () {
-    const items = this.items.map(({ avatar, id, name }) => {
-      return new LabelItem().render({ avatar, id, name })
-    })
+  renderArtists () {
+    const self = this
+
+    let items = clone(this.local.items)
+
+    if (this.local.shuffle) {
+      items = items.sort(() => Math.random() - 0.5)
+    }
+
     return html`
-      <ul class="labels list ma0 pa0 cf">
-        ${items}
+      <ul class="artists list ma0 pa0 cf">
+        ${items.map(artistItem)}
       </ul>
     `
+
+    function artistItem (props) {
+      const { id } = props
+      const artist = self.state.cache(Artist, `artist-item-${id}`)
+      return artist.render(props)
+    }
   }
 
   renderError () {
     return html`
       <div class="flex flex-column flex-auto w-100 items-center justify-center">
-        <p>Failed to fetch labels</p>
+        <p>Failed to fetch artists</p>
+        <div>
+          <button onclick=${() => this.emit('labels:reload', this.state.params.id)}>Try again</button>
+        </div>
       </div>
     `
   }
@@ -157,7 +127,7 @@ class Labels extends Component {
   renderPlaceholder () {
     return html`
       <div class="flex flex-column flex-auto w-100 items-center justify-center">
-        <p class="tc">No labels found</p>
+        <p class="tc">👽 No artists found</p>
       </div>
     `
   }
@@ -168,6 +138,7 @@ class Labels extends Component {
       count: 3,
       options: { animate: true, repeat: true, reach: 9, fps: 10 }
     })
+
     return html`
       <div class="flex flex-column flex-auto items-center justify-center h5">
         ${loader}
@@ -177,10 +148,10 @@ class Labels extends Component {
 
   update (props) {
     if (props) {
-      return compare(props.items, this.items)
+      return compare(props.items, this.local.items)
     }
     return false
   }
 }
 
-module.exports = Labels
+module.exports = Artists
