+++ beta/components/artists/random-grid.js
