+++ beta/components/artists/single.js
