+++ beta/components/favorite-button/index.js
