+++ beta/components/menu/index.js
@@ -3,8 +3,8 @@ const css = require('sheetify')
 const Component = require('choo/component')
 const nanostate = require('nanostate')
 const icon = require('@resonate/icon-element')
-const Search = require('./search')
-const matchMedia = require('../lib/match-media')
+const Search = require('../search')
+const matchMedia = require('../../lib/match-media')
 const Nanobounce = require('nanobounce')
 const nanobounce = Nanobounce()
 const ResizeObserver = require('resize-observer-polyfill')
