+++ beta/components/payment-methods/index.js
