+++ beta/components/profile-header/image.js
