+++ beta/components/search/index.js
