+++ beta/components/theme-switcher/index.js
