+++ beta/views/404/index.js
