+++ beta/views/artists/show.js
@@ -3,7 +3,7 @@ const html = require('choo/html')
 const Albums = require('../../components/albums')
 const Playlist = require('@resonate/playlist-component')
 const ProfileHeader = require('../../components/profile-header')
-const ProfileHeaderImage = require('../../components/profile-header-image')
+const ProfileHeaderImage = require('../../components/profile-header/image')
 const Labels = require('../../components/labels')
 const Links = require('../../components/links')
 const viewLayout = require('../../elements/view-layout')
